+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Reformatted autocomplete results
 - Trimmed down autocomplete results to display top 15 results, instead of hundreds
 - Hover effects over buttons for more responsiveness
+- mantine component library components
+- frontend is switched to vite+react, recommended by mantine library
 
 ### Updated
 - package-lock.json: Lock file was updated when ran with `npm i`, removal of testing libraries for react suggest a change in react's dependencies.